fix: refactor to use axios lib
